project-asistencias/contenedorAsistApp/requirements.txt
Django==3.0.8
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
django-tables2==2.3.1
mysqlclient==2.0.1
PyMySQL==0.9.3
requests
gunicorn==20.0.4